04.04.2024
Update Kernel

+++ Kernel/kernel/definition/basic/buffer.hpp
@@ -9,7 +9,7 @@ namespace Sen::Kernel::Definition
     concept CharacterOnView = std::is_same_v<Type, char> || std::is_same_v<Type, unsigned char>;
 
     template <typename Type>
-    concept IsValidArgument = std::is_same<Type, std::uint64_t>::value && true;
+    concept IsValidArgument = std::is_same<Type, std::size_t>::value && true;
 
     namespace Buffer
     {
